Add all other planets

+++ Assets/Scripts/RotateAround.cs
@@ -6,7 +6,7 @@ public class RotateAround : MonoBehaviour {
     [Tooltip("This is the object that the script's game object will rotate around")]
 	public Transform target; // the object to rotate around
     [Tooltip("This is the speed at which the object rotates")]
-	public int speed; // the speed of rotation
+	public float speed; // the speed of rotation
 	
 	void Start() {
 		if (target == null)
